ORD-758 move tags, remove cucumber dependencies
